[arm64][debugger] Fix bug with Single step
As per the ARMv8 manual D2.12.3, PSTATE.SS must
also be set to enable single step debugging for
a given process.

Additionally the OSLock must also be disabled as
per the ARMv8 manual D2.4.1: "If either the OS
lock or the OS double-lock is locked, debug
exceptions other than Breakpoint Instruction
exceptions are disabled."

So to fascilitate debugging on HW we also unlock
the OS Lock upon booting.

Change-Id: I436a58658150b0b6179620a9d7129cfa7d4db9df