[hypervisor][arm64] Trap guest debug reg access
To enable debugging on ARM64 we have to unlock
the OS Lock upon boot. We don't want the guest OS
to alter the state of the OS lock so we trap any
accesses to the debug registers and implement
them as WI/RAZ.

This should work as both Zircon and Linux unlock
the debug lock immediately once upon boot
regardless of whether or not the debug lock
is implemented.

An guest OS that attempts to query for the
presense of the debug lock will find that it is
unimplemented since the OSLSR is implemented as
RAZ.

Change-Id: Ifa42f4fa2a50fc3569eac9a194c19a33059d7151